requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# Why-Bot/embedsite/embed_gen.py: 1
Flask == 1.1.2

# Why-Bot/cogs/search/minecraft.py: 10
# Why-Bot/cogs/search/search.py: 6
aiofiles == 0.6.0

# Why-Bot/cogs/search/minecraft.py: 9
# Why-Bot/cogs/search/search.py: 5
aiohttp == 3.7.4.post0

# Why-Bot/cogs/search/reddit.py: 3
asyncpraw == 7.5.0

# Why-Bot/cogs/leveling/leveling.py: 4
discordLevelingSystem == 1.0.2

# Why-Bot/cogs/moderation/moderation.py: 7
discord_webhook == 0.14.0

# Why-Bot/cogs/events/welcome.py: 3
# Why-Bot/cogs/leveling/leveling.py: 5
easy_pil == 0.1.0

# Why-Bot/cogs/voice_channel/music.py: 18
gTTS == 2.2.2

# Why-Bot/cogs/search/google.py: 5
# Why-Bot/cogs/search/search.py: 7
google_api_python_client == 2.29.0

# Why-Bot/cogs/voice_channel/music.py: 19
mutagen == 1.45.1

# Why-Bot/cogs/games/counting.py: 7
# Why-Bot/cogs/utilities/utilities.py: 8
numexpr == 2.8.1

# Why-Bot/cogs/search/search.py: 12
praw == 7.5.0

# Why-Bot/cogs/utilities/info.py: 9
psutil == 5.9.0

# Why-Bot/cogs/economy/economy.py: 1,3
# Why-Bot/cogs/events/errors.py: 1,3
# Why-Bot/cogs/events/event_logging.py: 1,3
# Why-Bot/cogs/events/guild.py: 1,3
# Why-Bot/cogs/events/onping.py: 1,3
# Why-Bot/cogs/events/welcome.py: 1,2
# Why-Bot/cogs/fun/fun.py: 1,4,6,11
# Why-Bot/cogs/fun/text.py: 1,5,6
# Why-Bot/cogs/games/counting.py: 1,2,3
# Why-Bot/cogs/games/tictactoe.py: 1,2
# Why-Bot/cogs/leveling/leveling.py: 1,3
# Why-Bot/cogs/moderation/moderation.py: 1,4
# Why-Bot/cogs/moderation/ticket.py: 4
# Why-Bot/cogs/owner/blacklist.py: 1,2
# Why-Bot/cogs/owner/cog_tools.py: 1,2
# Why-Bot/cogs/owner/dmreply.py: 1,3
# Why-Bot/cogs/owner/fusion.py: 1,4
# Why-Bot/cogs/owner/git.py: 1,2
# Why-Bot/cogs/search/google.py: 1,3
# Why-Bot/cogs/search/minecraft.py: 1,5
# Why-Bot/cogs/search/reddit.py: 1,2
# Why-Bot/cogs/search/search.py: 1,3
# Why-Bot/cogs/setup_help/help.py: 4
# Why-Bot/cogs/setup_help/settings.py: 1,2
# Why-Bot/cogs/slash_cmds/slash.py: 3,4,5,6
# Why-Bot/cogs/utilities/info.py: 1,8
# Why-Bot/cogs/utilities/utilities.py: 7
# Why-Bot/cogs/utilities/whosplaying.py: 1,2
# Why-Bot/cogs/voice_channel/music.py: 20,21
# Why-Bot/cogs/voice_channel/voice.py: 1,3
# Why-Bot/main.py: 12
# Why-Bot/utils/buttons.py: 3,4
# Why-Bot/utils/other.py: 1
py_cord == 2.0.0a4850+gd2878166

# Why-Bot/main.py: 14
pyfiglet == 0.8.post1

# Why-Bot/cogs/fun/fun.py: 7
# Why-Bot/cogs/moderation/moderation.py: 6
# Why-Bot/cogs/owner/fusion.py: 6
# Why-Bot/cogs/search/google.py: 10
# Why-Bot/cogs/search/minecraft.py: 7
# Why-Bot/cogs/search/reddit.py: 5
# Why-Bot/cogs/search/search.py: 15
# Why-Bot/main.py: 11
python_dotenv == 0.15.0

# Why-Bot/cogs/voice_channel/music.py: 4
pyttsx3 == 2.90

# Why-Bot/cogs/utilities/utilities.py: 5
qrcode == 7.3.1

# Why-Bot/cogs/moderation/moderation.py: 10
# Why-Bot/cogs/voice_channel/music.py: 9
requests == 2.25.1

# Why-Bot/main.py: 15
sidspackage == 0.0.4

# Why-Bot/cogs/utilities/utilities.py: 6
simpcalci == 0.0.2

# Why-Bot/cogs/voice_channel/music.py: 8
wget == 3.2

# Why-Bot/cogs/search/search.py: 14
wikipedia == 1.4.0

# Why-Bot/cogs/voice_channel/music.py: 12
youtube_dl == 2021.6.6

# Why-Bot/cogs/voice_channel/music.py: 13
youtube_search_python == 1.4.9
